design: 도서 검색페이지 반응형 css 수정
